Create scripts dir; delete excess example videos
